separate fullscreen logic in post viewer to hook
